pyproject.toml
[project]
name = "compress-picture-to70"
version = "0.1.0"
description = "批量圖片壓縮工具，支援並行處理與 EXIF 保留"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "rich>=13.0.0",
]

[project.scripts]
compress-img = "compress_images:main"
images-to-webp = "images_to_webp:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
